feat: Connect to geotiff.js web worker decoder Pool (#61)
### Change list

- Add `pool` prop into `COGLayer` and `GeoTIFFLayer` to let users pass
in a custom geotiff.js decoder pool, which decodes image chunks on a web
worker instead of on the main thread.
- If a pool is not passed by a user, then we create a default worker
pool in global scope and share it for all instances of these layers.
- In theory, we could allow something like `false` for the prop if the
user explicitly didn't want to create web workers, but that can be for
the future if anyone asks for it.

Closes https://github.com/developmentseed/deck.gl-raster/issues/53